response_generation/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
numpy
scikit-learn
textstat
rapidfuzz
